feat: 添加用户认证系统和API令牌管理
- 新增用户模型和认证API
- 添加Google OAuth和GitHub OAuth配置
- 实现API令牌管理功能
- 添加用户系统数据库迁移脚本
- 更新项目和任务API以支持用户认证
- 添加环境配置示例文件
- 新增用户系统初始化脚本

+++ backend/models/attachment.py
@@ -15,7 +15,7 @@ class Attachment(db.Model):
     __tablename__ = 'attachments'
     
     id = Column(Integer, primary_key=True, autoincrement=True, comment='主键ID')
-    task_id = Column(Integer, ForeignKey('tasks.id'), nullable=False, comment='任务ID')
+    task_id = Column(BigInteger, ForeignKey('tasks.id'), nullable=False, comment='任务ID')
     filename = Column(String(255), nullable=False, comment='文件名')
     original_filename = Column(String(255), nullable=False, comment='原始文件名')
     file_path = Column(String(500), nullable=False, comment='文件路径')
